Added mongo token from .env to mongodb.go
